fix autoloading StringableObjectStub when running vendor/bin/phpunit ./tests/Support/SupportStringableTest.php directly
